add buisnes logic for comments and rating

+++ YourPlace.Web/YourPlace.Web/Models/Store/ListStoreViewModel.cs
@@ -34,7 +34,7 @@ namespace YourPlace.Web.Models.Store
         [Required]
         public string District { get; set; }
 
-        public int? Raiting { get; set; }
+        public double? Raiting { get; set; }
 
         public ICollection<StoreServices> StoreServices { get; set; } = new List<StoreServices>();
     }
